refactor : Post -> PostInfo -> PostResponse로 흐르는 데이터 흐름을 좀 더 직관적으로 표현
기존에 Application -> Presentation 데이터 이동에 Response를 사용하는 부분 수정
