--- src/main/java/com/sparksInTheStep/webBoard/content/service/dto/PostInfo.java
@@ -1,7 +1,0 @@
-package com.sparksInTheStep.webBoard.content.service.dto;
-
-public record PostInfo(String title){
-    public static PostInfo of(String title){
-        return new PostInfo(title);
-    }
-}
